requirements.txt
dtaidistance==2.3.13
h5py==3.14.0
matplotlib==3.9.4
mne==1.8.0
numpy==2.0.2
pandas==2.3.3
pretty_midi==0.2.11
PyYAML==6.0.3
scikit_learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
torch==1.13.1
tqdm==4.67.1
